+++ src/main/java/com/contentful/java/cda/CDAService.java
@@ -2,39 +2,32 @@ package com.contentful.java.cda;
 
 import java.util.Map;
 
-import retrofit2.Call;
+import io.reactivex.Flowable;
 import retrofit2.Response;
 import retrofit2.http.GET;
 import retrofit2.http.Path;
 import retrofit2.http.Query;
 import retrofit2.http.QueryMap;
-import rx.Observable;
 
 interface CDAService {
   @GET("spaces/{space}")
-  Observable<Response<CDASpace>> space(
+  Flowable<Response<CDASpace>> space(
       @Path("space") String space);
 
   @GET("spaces/{space}/{type}/{identifier}")
-  Observable<Response<CDAResource>> one(
-      @Path("space") String space,
-      @Path("type") String type,
-      @Path("identifier") String identifier);
-
-  @GET("spaces/{space}/{type}/{identifier}")
-  Call<CDAResource> oneAsCall(
+  Flowable<Response<CDAResource>> one(
       @Path("space") String space,
       @Path("type") String type,
       @Path("identifier") String identifier);
 
   @GET("spaces/{space}/{type}")
-  Observable<Response<CDAArray>> array(
+  Flowable<Response<CDAArray>> array(
       @Path("space") String space,
       @Path("type") String type,
       @QueryMap Map<String, String> query);
 
   @GET("spaces/{space}/sync")
-  Observable<Response<SynchronizedSpace>> sync(
+  Flowable<Response<SynchronizedSpace>> sync(
       @Path("space") String space,
       @Query("initial") Boolean initial,
       @Query("sync_token") String sync_token);