+++ src/main/java/com/contentful/java/cda/ObserveQuery.java
@@ -1,9 +1,9 @@
 package com.contentful.java.cda;
 
+import io.reactivex.Flowable;
+import io.reactivex.Scheduler;
+import io.reactivex.functions.Function;
 import retrofit2.Response;
-import rx.Observable;
-import rx.Scheduler;
-import rx.functions.Func1;
 
 import static com.contentful.java.cda.CDAType.ASSET;
 import static com.contentful.java.cda.CDAType.CONTENTTYPE;
@@ -11,11 +11,11 @@ import static com.contentful.java.cda.CDAType.ENTRY;
 import static com.contentful.java.cda.Util.typeForClass;
 
 /**
- * Represents a query to the Delivery API which may be invoked via an {@link Observable}
+ * Represents a query to the Delivery API which may be invoked via an {@link Flowable}
  * subscription.
- *
+ * <p>
  * Observable requests are subscribed and observed on the same thread that executed
- * the request. Call {@link Observable#subscribeOn(Scheduler)} and {@link Observable#observeOn(Scheduler)}
+ * the request. Call {@link Flowable#subscribeOn(Scheduler)} and {@link Flowable#observeOn(Scheduler)}
  * to control that.
  */
 public class ObserveQuery<T extends CDAResource> extends AbsQuery<T, ObserveQuery<T>> {
@@ -25,14 +25,16 @@ public class ObserveQuery<T extends CDAResource> extends AbsQuery<T, ObserveQuer
 
   /**
    * Observe a resource matching the given {@code id}.
+   *
    * @param id resource id.
-   * @return {@link Observable} instance.
+   * @return {@link Flowable} instance.
+   * @throws CDAResourceNotFoundException if resource was not found.
    */
-  public Observable<T> one(final String id) {
-    Observable<T> observable = where("sys.id", id).all().map(new Func1<CDAArray, T>() {
-      @Override @SuppressWarnings("unchecked") public T call(CDAArray array) {
+  public Flowable<T> one(final String id) {
+    Flowable<T> flowable = where("sys.id", id).all().map(new Function<CDAArray, T>() {
+      @Override @SuppressWarnings("unchecked") public T apply(CDAArray array) {
         if (array.items().size() == 0) {
-          return null;
+          throw new CDAResourceNotFoundException(type, id);
         }
         CDAType resourceType = typeForClass(type);
         if (ASSET.equals(resourceType)) {
@@ -48,8 +50,8 @@ public class ObserveQuery<T extends CDAResource> extends AbsQuery<T, ObserveQuer
     });
 
     if (CONTENTTYPE.equals(typeForClass(type))) {
-      observable = observable.map(new Func1<T, T>() {
-        @Override public T call(T t) {
+      flowable = flowable.map(new Function<T, T>() {
+        @Override public T apply(T t) {
           if (t != null) {
             client.cache.types().put(t.id(), (CDAContentType) t);
           }
@@ -57,21 +59,22 @@ public class ObserveQuery<T extends CDAResource> extends AbsQuery<T, ObserveQuer
         }
       });
     }
-    return observable;
+    return flowable;
   }
 
   /**
    * Observe an array of all resources matching the type of this query.
-   * @return {@link Observable} instance.
+   *
+   * @return {@link Flowable} instance.
    */
-  public Observable<CDAArray> all() {
+  public Flowable<CDAArray> all() {
     return client.cacheAll(false)
-        .flatMap(new Func1<Cache, Observable<Response<CDAArray>>>() {
-          @Override public Observable<Response<CDAArray>> call(Cache cache) {
+        .flatMap(new Function<Cache, Flowable<Response<CDAArray>>>() {
+          @Override public Flowable<Response<CDAArray>> apply(Cache cache) {
             return client.service.array(client.spaceId, path(), params);
           }
-        }).map(new Func1<Response<CDAArray>, CDAArray>() {
-          @Override public CDAArray call(Response<CDAArray> response) {
+        }).map(new Function<Response<CDAArray>, CDAArray>() {
+          @Override public CDAArray apply(Response<CDAArray> response) {
             return ResourceFactory.array(response, client);
           }
         });