+++ src/test/java/com/contentful/java/cda/EntryTest.java
@@ -6,34 +6,61 @@ import com.contentful.java.cda.lib.TestCallback;
 import org.junit.Test;
 
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
 
-import rx.functions.Action1;
+import io.reactivex.functions.Consumer;
 
 import static com.google.common.truth.Truth.assertThat;
 
 public class EntryTest extends BaseTest {
-  @Test
+  @Test(expected = CDAResourceNotFoundException.class)
   @Enqueue("array_empty.json")
   public void fetchNonExistingReturnsNull() throws Exception {
-    assertThat(client.fetch(CDAEntry.class).one("foo")).isNull();
+    try {
+      client.fetch(CDAEntry.class).one("foo");
+    } catch (NullPointerException e) {
+      assertThat(e.getMessage()).isEqualTo("Could not find id 'foo' of type 'CDAEntry'.");
+      throw e;
+    }
   }
 
   @Test
   @Enqueue("array_empty.json")
-  public void fetchNonExistingEntryEmitsNull() throws Exception {
+  public void fetchNonExistingEntryThrowsError() throws Exception {
     final Object result[] = new Object[]{new Object()};
+    final Object error[] = new Object[]{new Object()};
     final CountDownLatch latch = new CountDownLatch(1);
-    assertThat(client.observe(CDAEntry.class)
+
+    client.observe(CDAEntry.class)
         .one("foo")
-        .subscribe(new Action1<CDAEntry>() {
-          @Override public void call(CDAEntry entry) {
-            result[0] = entry;
-            latch.countDown();
-          }
-        }));
-    latch.await();
+        .subscribe(
+            new Consumer<CDAEntry>() {
+              @Override public void accept(CDAEntry entry) {
+                result[0] = entry;
+                error[0] = null;
+
+                latch.countDown();
+              }
+            }, new Consumer<Throwable>() {
+              @Override public void accept(Throwable throwable) {
+                result[0] = null;
+                error[0] = throwable;
+
+                latch.countDown();
+              }
+            }
+        );
+
+    latch.await(1, TimeUnit.SECONDS);
+
     assertThat(result[0]).isNull();
+    assertThat(error[0]).isNotNull();
+    assertThat(error[0]).isInstanceOf(CDAResourceNotFoundException.class);
+    final CDAResourceNotFoundException notFoundException = (CDAResourceNotFoundException) error[0];
+    final String message = notFoundException.getMessage();
+    assertThat(message).isEqualTo("Could not find id 'foo' of type 'CDAEntry'.");
   }
 
   @Test
@@ -43,7 +70,10 @@ public class EntryTest extends BaseTest {
         .one("foo", new TestCallback<CDAEntry>())
         .await();
 
-    assertThat(callback.error()).isNull();
+    assertThat(callback.error()).isNotNull();
+    final String message = callback.error().getMessage();
+    assertThat(message).isEqualTo("Could not find id 'foo' of type 'CDAEntry'.");
+
     assertThat(callback.result()).isNull();
   }
 
@@ -85,9 +115,107 @@ public class EntryTest extends BaseTest {
       }
     });
 
-    latch.await();
+    latch.await(1, TimeUnit.SECONDS);
     assertThat(result[0]).isNotNull();
     assertNyanCat(result[0]);
+  }
+
+  @Test
+  @Enqueue(
+      defaults = {
+          "demo/space.json",
+          "content_types/populate_cache_simple.json"
+      },
+      value = {
+          "content_types/populate_cache_simple.json"
+      }
+  )
+  public void populateAllContentTypesSinglePage() throws Exception {
+    final int contentTypesCached = client.populateContentTypeCache();
+
+    assertThat(contentTypesCached).isEqualTo(3);
+
+    final Map<String, CDAContentType> types = client.cache.types();
+    assertThat(types.size()).isEqualTo(3);
+
+    final CDAContentType first = types.get("001");
+    assertThat(first.fields.size()).isEqualTo(3);
+    assertThat(first.fields.get(0).id).isEqualTo("first");
+    assertThat(first.fields.get(0).type).isEqualTo("Symbol");
+
+    final CDAContentType second = types.get("002");
+    assertThat(second.fields.size()).isEqualTo(3);
+    assertThat(second.fields.get(0).id).isEqualTo("first");
+    assertThat(second.fields.get(0).type).isEqualTo("Symbol");
+
+    final CDAContentType third = types.get("003");
+    assertThat(third.fields.size()).isEqualTo(3);
+    assertThat(third.fields.get(0).id).isEqualTo("first");
+    assertThat(third.fields.get(0).type).isEqualTo("Symbol");
+  }
+
+  @Test
+  @Enqueue(
+      defaults = {
+          "demo/space.json",
+          "content_types/populate_cache_simple.json"
+      },
+      value = {
+          "content_types/populate_cache_complex_p1.json",
+          "content_types/populate_cache_complex_p2.json",
+          "content_types/populate_cache_complex_p3.json"
+      }
+  )
+  public void populateAllContentTypesMultiplePages() throws Exception {
+    int numberOfContentTypes = client.populateContentTypeCache(60);
+
+    assertThat(numberOfContentTypes).isEqualTo(151);
+    assertThat(client.cache.types().size()).isEqualTo(151);
+
+    final Map<String, CDAContentType> types = client.cache.types();
+    final CDAContentType first = types.get("001");
+    assertThat(first.fields.size()).isEqualTo(3);
+    assertThat(first.fields.get(0).id).isEqualTo("first");
+    assertThat(first.fields.get(0).type).isEqualTo("Symbol");
+
+    final CDAContentType second = types.get("079");
+    assertThat(second.fields.size()).isEqualTo(3);
+    assertThat(second.fields.get(0).id).isEqualTo("first");
+    assertThat(second.fields.get(0).type).isEqualTo("Symbol");
+
+    final CDAContentType third = types.get("151");
+    assertThat(third.fields.size()).isEqualTo(3);
+    assertThat(third.fields.get(0).id).isEqualTo("first");
+    assertThat(third.fields.get(0).type).isEqualTo("Symbol");
+  }
+
+  @Test
+  @Enqueue(
+      defaults = {
+          "demo/space.json",
+          "content_types/populate_cache_simple.json"
+      },
+      value = {
+          "content_types/populate_cache_complex_p1.json",
+          "content_types/populate_cache_complex_p2.json",
+          "content_types/populate_cache_complex_p3.json",
+          "content_types/populate_cache_last_entry.json"
+      }
+  )
+  public void aPopulatedContentTypeCacheDoesNotToFetchContentTypes() throws Exception {
+    client.populateContentTypeCache(60);
+
+    assertThat(client.cache.types().size()).isEqualTo(151);
+
+    final CDAContentType lastContentType = client.cache.types().get("151");
+    assertThat(lastContentType.fields.size()).isEqualTo(3);
+    assertThat(lastContentType.fields.get(0).id).isEqualTo("first");
+    assertThat(lastContentType.fields.get(0).type).isEqualTo("Symbol");
+
+    final CDAEntry lastEntry = client.fetch(CDAEntry.class).one("151");
+
+    assertThat(lastEntry).isNotNull();
+    assertThat(lastEntry.contentType().id()).isEqualTo(lastContentType.id());
   }
 
   @Test